refactor(shoppingcart): move all customer logic to auth package
