+++ src/main/java/woowacourse/auth/dto/request/TokenRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.dto;
+package woowacourse.auth.dto.request;
 
 import javax.validation.constraints.NotNull;
