+++ src/main/java/woowacourse/auth/dto/response/GetMeResponse.java
@@ -1,7 +1,7 @@
-package woowacourse.user.dto.response;
+package woowacourse.auth.dto.response;
 
 import java.util.Objects;
-import woowacourse.user.domain.Customer;
+import woowacourse.auth.domain.Customer;
 
 public class GetMeResponse {
