+++ src/main/java/woowacourse/auth/dto/response/TokenResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.dto;
+package woowacourse.auth.dto.response;
 
 public class TokenResponse {
