+++ src/main/java/woowacourse/auth/ui/AuthController.java
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import woowacourse.auth.application.AuthService;
-import woowacourse.auth.dto.TokenRequest;
-import woowacourse.auth.dto.TokenResponse;
+import woowacourse.auth.dto.request.TokenRequest;
+import woowacourse.auth.dto.response.TokenResponse;
 
 @RestController
 public class AuthController {
