+++ src/test/java/woowacourse/auth/application/AuthServiceTest.java
@@ -15,8 +15,8 @@ import woowacourse.auth.domain.EncryptedPassword;
 import woowacourse.auth.domain.Password;
 import woowacourse.auth.domain.Token;
 import woowacourse.auth.domain.User;
-import woowacourse.auth.dto.TokenRequest;
-import woowacourse.auth.dto.TokenResponse;
+import woowacourse.auth.dto.request.TokenRequest;
+import woowacourse.auth.dto.response.TokenResponse;
 import woowacourse.common.exception.AuthenticationException;
 import woowacourse.util.DatabaseFixture;
